refactor: move discount-codes plugin to Node app
