+++ imports/node-app/core-services/discounts/queries/index.js
@@ -1,0 +1,5 @@
+import getDiscountsTotalForCart from "./getDiscountsTotalForCart.js";
+
+export default {
+  getDiscountsTotalForCart
+};
