+++ imports/node-app/plugins/discount-codes/i18n/fr.json
