+++ imports/node-app/plugins/discount-codes/startup.js
@@ -1,13 +1,11 @@
-import appEvents from "/imports/node-app/core/util/appEvents";
-
 /**
  * @summary Called on startup
  * @param {Object} context Startup context
  * @param {Object} context.collections Map of MongoDB collections
  * @returns {undefined}
  */
-export default function startup({ collections }) {
-  const { Discounts } = collections;
+export default function startup(context) {
+  const { appEvents, collections: { Discounts } } = context;
 
   appEvents.on("afterOrderCreate", async ({ order }) => {
     if (Array.isArray(order.discounts)) {
