+++ imports/node-app/registerPlugins.js
@@ -17,6 +17,7 @@ import registerTaxesPlugin from "./core-services/taxes/index.js";
 /* plugins */
 import registerAddressPlugin from "./plugins/address/index.js";
 import registerDashboardPlugin from "./plugins/dashboard/index.js";
+import registerDiscountCodesPlugin from "./plugins/discount-codes/index.js";
 import registerEmailTemplatesPlugin from "./plugins/email-templates/index.js";
 import registerExamplePaymentsPlugin from "./plugins/payments-example/index.js";
 import registerNavigationPlugin from "./plugins/navigation/index.js";
@@ -42,7 +43,6 @@ import registerAccountsPlugin from "/imports/plugins/core/accounts/server/no-met
 import registerCartPlugin from "/imports/plugins/core/cart/server/no-meteor/register";
 import registerCheckoutPlugin from "/imports/plugins/core/checkout/server/no-meteor/register";
 import registerCorePlugin from "/imports/plugins/core/core/server/no-meteor/register";
-import registerDiscountCodesPlugin from "/imports/plugins/included/discount-codes/server/no-meteor/register";
 import registerJobQueuePlugin from "/imports/plugins/included/job-queue/server/no-meteor/register";
 import registerMarketplacePlugin from "/imports/plugins/included/marketplace/server/no-meteor/register";
 import registerOrdersPlugin from "/imports/plugins/core/orders/server/no-meteor/register";
