+++ imports/plugins/core/orders/server/no-meteor/mutations/placeOrder.js
@@ -6,7 +6,6 @@ import ReactionError from "@reactioncommerce/reaction-error";
 import { getAnonymousAccessToken } from "../util/anonymousToken";
 import appEvents from "/imports/node-app/core/util/appEvents";
 import { Order as OrderSchema, Payment as PaymentSchema } from "/imports/collections/schemas";
-import getDiscountsTotalForCart from "/imports/plugins/core/discounts/server/no-meteor/util/getDiscountsTotalForCart";
 import { getPaymentMethodConfigByName } from "/imports/node-app/core-services/payments/registration.js"; // TODO: remove cross-plugin import (https://github.com/reactioncommerce/reaction/issues/5653)
 import buildOrderFulfillmentGroupFromInput from "../util/buildOrderFulfillmentGroupFromInput";
 import verifyPaymentsMatchOrderTotal from "../../util/verifyPaymentsMatchOrderTotal";
@@ -191,7 +190,7 @@ export default async function placeOrder(context, input) {
   let discounts = [];
   let discountTotal = 0;
   if (cart) {
-    const discountsResult = await getDiscountsTotalForCart(context, cart);
+    const discountsResult = await context.queries.getDiscountsTotalForCart(context, cart);
     ({ discounts } = discountsResult);
     discountTotal = discountsResult.total;
   }
